% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{bnlearnPlot}
\alias{bnlearnPlot}
\title{Plots bnlearn graph model with DiagrammeR}
\usage{
bnlearnPlot(model, labels = NULL)
}
\arguments{
\item{model}{A model fit object of class bn.fit}

\item{labels}{An optional named list of variable labels}
}
\value{
A Diagrammer plot of the path diagram for \code{model}
}
\description{
Plots bnlearn graph model with DiagrammeR
}

